+++ packages/angular-eslint/src/configs/template-all.ts
@@ -40,6 +40,7 @@ export default (
       '@angular-eslint/template/no-negated-async': 'error',
       '@angular-eslint/template/no-nested-tags': 'error',
       '@angular-eslint/template/no-positive-tabindex': 'error',
+      '@angular-eslint/template/prefer-at-else': 'error',
       '@angular-eslint/template/prefer-at-empty': 'error',
       '@angular-eslint/template/prefer-built-in-pipes': 'error',
       '@angular-eslint/template/prefer-contextual-for-variables': 'error',