+++ packages/bundled-angular-compiler/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 20.4.0 (2025-10-12)
+
+This was a version bump only for bundled-angular-compiler to align it with other projects, there were no code changes.
+
+## 20.3.0 (2025-09-14)
+
+This was a version bump only for bundled-angular-compiler to align it with other projects, there were no code changes.
+
+## 20.2.0 (2025-08-22)
+
+This was a version bump only for bundled-angular-compiler to align it with other projects, there were no code changes.
+
 ## 20.1.1 (2025-06-19)
 
 This was a version bump only for bundled-angular-compiler to align it with other projects, there were no code changes.